+++ src/commands/MODEL/modelCommandParser.ts
@@ -5,10 +5,12 @@ import { MODEL_VARIANTS } from '../../types/ModelVariant';
 import type { PipelineJson } from '../../types/PipelineJson/PipelineJson';
 import type { string_markdown_text } from '../../types/typeAliases';
 import { keepUnused } from '../../utils/organization/keepUnused';
-import type { $PipelineJson } from '../_common/types/CommandParser';
-import type { $TemplateJson } from '../_common/types/CommandParser';
-import type { CommandParserInput } from '../_common/types/CommandParser';
-import type { PipelineBothCommandParser } from '../_common/types/CommandParser';
+import type {
+    $PipelineJson,
+    $TemplateJson,
+    CommandParserInput,
+    PipelineBothCommandParser,
+} from '../_common/types/CommandParser';
 import type { ModelCommand } from './ModelCommand';
 
 /**
@@ -130,7 +132,7 @@ export const modelCommandParser: PipelineBothCommandParser<ModelCommand> = {
         $templateJson: $TemplateJson,
         // $pipelineJson: $PipelineJson,
     ): void {
-        if ($templateJson.blockType !== 'PROMPT_TEMPLATE') {
+        if ($templateJson.templateType !== 'PROMPT_TEMPLATE') {
             throw new ParseError(`MODEL command can only be used in PROMPT_TEMPLATE block`);
         }