+++ src/commands/PERSONA/personaCommandParser.ts
@@ -4,10 +4,12 @@ import { ParseError } from '../../errors/ParseError';
 import type { PipelineJson } from '../../types/PipelineJson/PipelineJson';
 import type { string_markdown_text } from '../../types/typeAliases';
 import { keepUnused } from '../../utils/organization/keepUnused';
-import type { $PipelineJson } from '../_common/types/CommandParser';
-import type { $TemplateJson } from '../_common/types/CommandParser';
-import type { CommandParserInput } from '../_common/types/CommandParser';
-import type { PipelineBothCommandParser } from '../_common/types/CommandParser';
+import type {
+    $PipelineJson,
+    $TemplateJson,
+    CommandParserInput,
+    PipelineBothCommandParser,
+} from '../_common/types/CommandParser';
 import type { PersonaCommand } from './PersonaCommand';
 
 /**
@@ -130,7 +132,7 @@ function $applyToTemplateJson(
     const { personaName, personaDescription } = command;
 
     if ($templateJson !== null) {
-        if ($templateJson.blockType !== 'PROMPT_TEMPLATE') {
+        if ($templateJson.templateType !== 'PROMPT_TEMPLATE') {
             throw new ParseError(`PERSONA command can be used only in PROMPT_TEMPLATE block`);
         }